- move v1 endpoints into v1 folder
