+++ auth-api/src/auth_api/resources/v1/entity.py
