+++ auth-api/src/auth_api/resources/v1/permissions.py
