+++ auth-api/src/auth_api/resources/v1/products.py
